- added flow manager - redesigned slot senders to use VoiceBitmap<32> - redesigned entire listener system - implemented hashed paths & redesigned suspend system

+++ hi_core/hi_core/MacroControlledComponents.cpp
@@ -433,7 +433,7 @@ void MacroControlledObject::setup(Processor *p, int parameter_, const String &na
 	parameter = parameter_;
 	name = name_;
 
-	auto a = (uint8)parameter;
+	auto a = (uint16)parameter;
 
 	p->addAttributeListener(valueListener, &a, 1, dispatch::DispatchType::sendNotificationAsync);
